+++ altgraph/GraphUtil.py
@@ -6,6 +6,7 @@ altgraph.GraphUtil - Utility classes and functions
 import random
 from collections import deque
 from altgraph import Graph
+from altgraph import GraphError
 
 def generate_random_graph(node_num, edge_num, self_loops=False, multi_edges=False):
     '''
@@ -13,6 +14,16 @@ def generate_random_graph(node_num, edge_num, self_loops=False, multi_edges=Fals
     randomly connected by *edge_num* edges.
     '''
     g = Graph.Graph()
+
+    if not multi_edges:
+        if self_loops:
+            max_edges = node_num * node_num
+        else:
+            max_edges = node_num * (node_num-1)
+
+        if edge_num > max_edges:
+            raise GraphError("inconsistent arguments to 'generate_random_graph'")
+
     nodes = range(node_num)
 
     for node in nodes:
@@ -39,22 +50,25 @@ def generate_random_graph(node_num, edge_num, self_loops=False, multi_edges=Fals
 
 def generate_scale_free_graph(steps, growth_num, self_loops=False, multi_edges=False):
     '''
-    Generates and returns a :py:class:`~altgraph.Graph.Graph` instance that will have *steps*growth_num* nodes
+    Generates and returns a :py:class:`~altgraph.Graph.Graph` instance that will have *steps* \* *growth_num* nodes
     and a scale free (powerlaw) connectivity. Starting with a fully connected graph with *growth_num* nodes
     at every step *growth_num* nodes are added to the graph and are connected to existing nodes with
     a probability proportional to the degree of these existing nodes.
     '''
+    # FIXME: The code doesn't seem to do what the documentation claims.
     graph = Graph.Graph()
 
     # initialize the graph
     store = []
     for i in range(growth_num):
-        store   += [ i ] * (growth_num - 1)
+        #store   += [ i ] * (growth_num - 1)
         for j in range(i + 1, growth_num):
+            store.append(i)
+            store.append(j)
             graph.add_edge(i,j)
 
     # generate
-    for node in range(growth_num, (steps-1) * growth_num):
+    for node in range(growth_num, steps * growth_num):
         graph.add_node(node)
         while ( graph.out_degree(node) < growth_num ):
             nbr = random.choice(store)